canvas: Move peniko/kurbo conversions in separate file
